Move some actions to OrderServices
