+++ app/Http/Controllers/OrderController.php
@@ -5,21 +5,10 @@ namespace App\Http\Controllers;
 use App\Models\Order;
 use App\Models\Bike;
 use App\Http\Requests\CreateOrderRequest;
-use Illuminate\Support\Facades\Auth;
+use App\Services\OrderServices;
 
 class OrderController extends Controller
 {
-    /**
-     * Error messages.
-     * 
-     * @var array
-     */
-    private $errorMessages = [
-        'out-of-stock' => [
-            'Mặt hàng :item không đủ, chỉ còn :stock trong kho.'
-        ],
-    ];
-
     /**
      * Success messages.
      * 
@@ -38,68 +27,11 @@ class OrderController extends Controller
     ];
 
     /**
-     * Get error message telling that item is out of stock.
+     * Order Service will be using.
      * 
-     * @param  \App\Models\Bike $bike
-     * @return string
+     * @var  \App\Services\OrderServices
      */
-    private function getItemOutOfStockError(Bike $item) {
-        $itemError = str_replace(
-            ':item', 
-            $item->bike_name,
-            $this->errorMessages['out-of-stock']
-        );
-        
-        return str_replace(
-            ':stock',
-            $item->bike_stock,
-            $itemError
-        );
-    }
-
-    /**
-     * Validate item quantity on Create.
-     * aka checking if that item is out of stock or not.
-     * 
-     * @param  array $validator
-     * @return array
-     */
-    private function validateItemQuantityCreate($validator) {
-        $errors = [];
-
-        foreach ($validator['order_detail'] as $order_detail) {
-            $bike = Bike::find($order_detail['bike_id']);
-            $order_value = (int)$order_detail['order_value'];
-
-            if ($bike->bike_stock < $order_value) {
-                array_push($errors, $this->getItemOutOfStockError($bike));
-            }
-        }
-
-        return $errors;
-    }
-
-    /**
-     * Validate item quantity on Update.
-     * 
-     * @param  array $validator
-     * @param  \App\Models\Order $order
-     * @return array
-     */
-    private function validateItemQuantityUpdate($validator, Order $order) {
-        $errors = [];
-
-        foreach ($validator['order_detail'] as $order_detail) {
-            $bike = Bike::find($order_detail['bike_id']);
-            $order_value = (int)$order_detail['order_value'];
-
-            if ($bike->bike_stock + $order->orderValue($bike) < $order_value) {
-                array_push($errors, $this->getItemOutOfStockError($bike));
-            }
-        }
-
-        return $errors;
-    }
+    private $orderServices;
 
     /**
      * Constructor for OrderController.
@@ -108,6 +40,7 @@ class OrderController extends Controller
      */
     public function __construct() {
         $this->authorizeResource(Order::class, 'order');
+        $this->orderServices = new OrderServices;
     }
 
     /**
@@ -138,8 +71,10 @@ class OrderController extends Controller
      */
     public function store(CreateOrderRequest $request) {
         $validator = $request->validated();
+        $validator['order_checkout'] = $request->has('order_checkout');
 
-        $quantityErrors = $this->validateItemQuantityCreate($validator);
+        $quantityErrors = $this->orderServices
+            ->validateItemQuantityCreate($validator);
 
         if (count($quantityErrors) > 0) {
             return redirect()
@@ -148,26 +83,7 @@ class OrderController extends Controller
                 ->withErrors($quantityErrors);
         }
 
-        $new_order = Order::create([
-            'customer_name' => $validator['customer_name'],
-            'customer_email' => $validator['customer_email'],
-            'checkout_at' => $request->has('order_checkout')
-                ? \Carbon\Carbon::now()->toDateTimeString()
-                : NULL
-        ]);
-
-        foreach ($validator['order_detail'] as $order_detail) {
-            $bike = Bike::find($order_detail['bike_id']);
-
-            $bike->bike_stock -= (int)$order_detail['order_value'];
-            $bike->save();
-
-            $new_order->bikes()->attach($bike->id, [
-                'order_value' => $order_detail['order_value'],
-                'order_buy_price' => $bike->bike_buy_price,
-                'order_sell_price' => $bike->bike_sell_price,
-            ]);
-        }
+        $new_order = $this->orderServices->createOrder($validator);
 
         return redirect()
             ->route('orders.show', $new_order->id)
@@ -214,8 +130,10 @@ class OrderController extends Controller
      */
     public function update(CreateOrderRequest $request, Order $order) {
         $validator = $request->validated();
+        $validator['order_checkout'] = $request->has('order_checkout');
 
-        $quantityErrors = $this->validateItemQuantityUpdate($validator, $order);
+        $quantityErrors = $this->orderServices
+            ->validateItemQuantityUpdate($validator, $order);
 
         if (count($quantityErrors) > 0) {
             return redirect()
@@ -224,41 +142,7 @@ class OrderController extends Controller
                 ->withErrors($quantityErrors);
         }
 
-        // Get new information of updated bikes.
-        $bikes_updated = [];
-
-        foreach ($validator['order_detail'] as $order_detail) {
-            $bike = Bike::find($order_detail['bike_id']);
-
-            $bike->bike_stock -= (int)$order_detail['order_value']
-                                    - $order->orderValue($bike);
-
-            $bike->save();
-
-            $bikes_updated[$bike->id] = [
-                'order_value' => $order_detail['order_value'],
-                'order_buy_price' => $bike->bike_buy_price,
-                'order_sell_price' => $bike->bike_sell_price
-            ];
-        }
-
-        // Restore bikes that was eliminated from the Order.
-        foreach ($order->bikes as $bike) {
-            if (!array_key_exists($bike->id, $bikes_updated)) {
-                $bike->bike_stock += $order->orderValue($bike);
-                $bike->save();
-            }
-        }
-
-        $order->bikes()->sync($bikes_updated);
-        $order->customer_name = $validator['customer_name'];
-        $order->customer_email = $validator['customer_email'];
-
-        $order->checkout_at = $request->has('order_checkout')
-            ? \Carbon\Carbon::now()->toDateTimeString()
-            : NULL;
-
-        $order->save();
+        $this->orderServices->updateOrder($order, $validator);
 
         return redirect()
             ->route('orders.show', $order->id)
